% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCNet}
\alias{PCNet}
\title{Adjacency matrix of the PCNet graph (v1.3)}
\format{:
  sparseMatrix with rows/columns named with gene symbols.}
\source{
\url{http://www.ndexbio.org/#/network/4de852d9-9908-11e9-bcaf-0ac135e8bacf}
}
\usage{
PCNet
}
\description{
Adjacency matrix of the PCNet graph (v1.3)
}
\keyword{datasets}
